feat(app): move to follow the "component folder pattern"
this is an mvp/1.0.0 version, I should probably turn this into a template I do it so often

BREAKING CHANGE: really just baseline build and ci/cd, tagging it breaking to trigger the 1.0.0 bump
