+++ src/index.tsx
@@ -3,7 +3,7 @@ import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import { App } from './App';
+import { App } from './components/App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
